+++ src/ts/platforms/apple-appstore/appstore-adapter.ts
@@ -185,7 +185,7 @@ namespace CdvPurchase {
                 this.pseudoReceipt.transactions = this.pseudoReceipt.transactions.filter(t => t.transactionId !== transactionId);
             }
 
-            /** Insert or update a transaction in the pseudo receipt */
+            /** Insert or update a transaction in the pseudo receipt, based on data collected from the native side */
             private async upsertTransaction(productId: string, transactionId: string, state: TransactionState): Promise<SKTransaction> {
                 return new Promise(resolve => {
                     this.initializeAppReceipt(() => {